requirements.txt
streamlit
google-generativeai
flask
flask-cors
requests
python-dotenv
